requirements.txt
Flask
openai
deepl
python-dotenv
flask
requests

